+++ okhttp/src/test/java/okhttp3/ConscryptTest.kt
@@ -22,18 +22,14 @@ import okhttp3.internal.platform.ConscryptPlatform
 import okhttp3.internal.platform.Platform
 import okhttp3.testing.PlatformRule
 import okhttp3.tls.internal.TlsUtil
-import okio.ByteString.Companion.toByteString
 import org.assertj.core.api.Assertions.assertThat
 import org.conscrypt.Conscrypt
-import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertFalse
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.RegisterExtension
-import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.ValueSource
 import javax.net.ssl.SSLSocket
 
 class ConscryptTest {
@@ -63,79 +59,6 @@ class ConscryptTest {
   @Test
   fun testTrustManager() {
     assertThat(Conscrypt.isConscrypt(Platform.get().platformTrustManager())).isTrue()
-  }
-
-  @ParameterizedTest(name = "{displayName}({arguments})")
-  @ValueSource(strings = ["TLSv1.2", "TLSv1.3"])
-  fun testSessionReuse(tlsVersion: String) {
-    val sessionIds = mutableListOf<String>()
-
-    enableTls()
-
-    val tlsVersion = TlsVersion.forJavaName(tlsVersion)
-    val spec = ConnectionSpec.Builder(ConnectionSpec.MODERN_TLS)
-      .tlsVersions(tlsVersion)
-      .build()
-
-    var reuseSession = false
-
-    val sslContext = handshakeCertificates.sslContext()
-    val systemSslSocketFactory = sslContext.socketFactory
-    val sslSocketFactory = object : DelegatingSSLSocketFactory(systemSslSocketFactory) {
-      override fun configureSocket(sslSocket: SSLSocket): SSLSocket {
-        return sslSocket.apply {
-          if (reuseSession) {
-            this.enableSessionCreation = false
-          }
-        }
-      }
-    }
-
-    client = client.newBuilder().connectionSpecs(listOf(spec))
-      .eventListenerFactory(clientTestRule.wrap(object : EventListener() {
-        override fun connectionAcquired(call: Call, connection: Connection) {
-          val sslSocket = connection.socket() as SSLSocket
-
-          sessionIds.add(sslSocket.session.id.toByteString().hex())
-        }
-      }))
-      .sslSocketFactory(sslSocketFactory, handshakeCertificates.trustManager)
-      .build()
-
-    server.enqueue(MockResponse().setBody("abc1"))
-    server.enqueue(MockResponse().setBody("abc2"))
-
-    val request = Request.Builder().url(server.url("/")).build()
-
-    client.newCall(request).execute().use { response ->
-      assertEquals(200, response.code)
-    }
-
-    client.connectionPool.evictAll()
-    assertEquals(0, client.connectionPool.connectionCount())
-
-    // Force reuse
-    reuseSession = true
-
-    client.newCall(request).execute().use { response ->
-      assertEquals(200, response.code)
-    }
-
-    assertEquals(2, sessionIds.size)
-    val directSessionIds =
-      sslContext.clientSessionContext.ids.toList().map { it.toByteString().hex() }
-    if (tlsVersion == TlsVersion.TLS_1_3) {
-      assertThat(sessionIds[0]).isBlank()
-      assertThat(sessionIds[1]).isBlank()
-
-      // https://github.com/google/conscrypt/issues/985
-      // assertThat(directSessionIds).containsExactlyInAnyOrder(sessionIds[0], sessionIds[1])
-    } else {
-      assertThat(sessionIds[0]).isNotBlank()
-      assertThat(sessionIds[1]).isNotBlank()
-
-      assertThat(directSessionIds).containsExactlyInAnyOrder(sessionIds[1])
-    }
   }
 
   @Test
